split components into files
