+++ src/components/CSVModal/QuerySelector.jsx
@@ -1,0 +1,36 @@
+import React from 'react';
+import './style.scss';
+import SelectionPill from './SelectionPill';
+
+function QuerySelector(props) {
+  const {
+    dataState, setDataState, queryParamType, allowMultiple,
+  } = props;
+
+  return (
+    <div className="selection-container pollutant">
+      <div className="selection-title-container">
+        <span className="selection-title">
+          Select
+          {' '}
+          {queryParamType}
+        </span>
+      </div>
+      <div className="horizontal-list-container">
+        {
+          dataState.map((type) => (
+            <SelectionPill
+              dataParameter={type}
+              state={dataState}
+              changeState={setDataState}
+              allowMultiple={allowMultiple}
+              index={0}
+            />
+          ))
+        }
+      </div>
+    </div>
+  );
+}
+
+export default QuerySelector;
